refactor codes according to PR comments.
