+++ azure-maven-plugin-lib/src/main/java/com/microsoft/azure/maven/telemetry/AppInsightHelper.java
@@ -7,7 +7,7 @@ package com.microsoft.azure.maven.telemetry;
 
 import com.microsoft.applicationinsights.TelemetryClient;
 import com.microsoft.applicationinsights.internal.channel.common.ApacheSenderFactory;
-import com.microsoft.azure.toolkit.lib.common.utils.HashMacUtils;
+import com.microsoft.azure.toolkit.lib.common.utils.InstallationIdUtils;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -28,7 +28,7 @@ public enum AppInsightHelper implements TelemetryProxy {
     AppInsightHelper() {
         sessionId = UUID.randomUUID().toString();
         defaultProperties.put(TELEMETRY_KEY_SESSION_ID, sessionId);
-        defaultProperties.put(TELEMETRY_KEY_INSTALLATIONID, HashMacUtils.getHashMac());
+        defaultProperties.put(TELEMETRY_KEY_INSTALLATIONID, InstallationIdUtils.getHashMac());
         initTelemetryHttpClient();
     }
 
@@ -80,7 +80,7 @@ public enum AppInsightHelper implements TelemetryProxy {
     }
 
     public String getInstallationId() {
-        return HashMacUtils.getHashMac();
+        return InstallationIdUtils.getHashMac();
     }
 
     public void addDefaultProperty(String key, String value) {
